Update user mutations namespace
