--- app/graphql/mutations/login_user_mutation.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Mutations
-  class LoginUserMutation < Mutations::BaseMutation
-    argument :input, Types::Custom::Inputs::Mutations::LoginUserInput, required: true
-    type Types::Custom::Objects::UserObject
-
-    def resolve(params)
-      super(params)
-      ::Users::HandleLoginUserService.call(params: @params, session: context.fetch(:session))
-    end
-  end
-end
