--- app/graphql/mutations/update_user_mutation.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Mutations
-  class UpdateUserMutation < Mutations::BaseMutation
-    argument :input, Types::Custom::Inputs::Mutations::UpdateUserInput, required: true
-    type Types::Custom::Objects::UserObject
-
-    def resolve(params)
-      super(params)
-      ::Users::UpdateUserService.call(params: @params)
-    end
-  end
-end
