+++ app/graphql/types/mutation_type.rb
@@ -3,15 +3,15 @@
 module Types
   class MutationType < Types::BaseObject
     field :register_user,
-          mutation: Mutations::RegisterUserMutation,
+          mutation: Mutations::Users::RegisterUserMutation,
           description: 'Register user'
 
     field :login_user,
-          mutation: Mutations::LoginUserMutation,
+          mutation: Mutations::Users::LoginUserMutation,
           description: 'Login user'
 
     field :logout_user,
-          mutation: Mutations::LogoutUserMutation,
+          mutation: Mutations::Users::LogoutUserMutation,
           description: 'Logout user'
 
     field :subscribe_user_to_newsletter,
@@ -23,11 +23,11 @@ module Types
           description: 'Remove user from newsletter'
 
     field :update_user,
-          mutation: Mutations::UpdateUserMutation,
+          mutation: Mutations::Users::UpdateUserMutation,
           description: 'Update user'
 
     field :remove_user,
-          mutation: Mutations::RemoveUserMutation,
+          mutation: Mutations::Users::RemoveUserMutation,
           description: "Remove user's account"
 
     field :add_opinion,
